+++ sdk/core/azure_core/src/http/policies/bearer_token_policy.rs
@@ -155,7 +155,7 @@ mod tests {
         Arc,
     };
     use typespec_client_core::{
-        http::{policies::TransportPolicy, BufResponse, Method, TransportOptions},
+        http::{policies::TransportPolicy, BufResponse, Method, Transport},
         time::Duration,
     };
 
@@ -210,9 +210,7 @@ mod tests {
         let credential = MockCredential::new(&[]);
         let policy = BearerTokenCredentialPolicy::new(Arc::new(credential), ["scope"]);
         let client = MockHttpClient::new(|_| panic!("expected an error from get_token"));
-        let transport = Arc::new(TransportPolicy::new(TransportOptions::new(Arc::new(
-            client,
-        ))));
+        let transport = Arc::new(TransportPolicy::new(Transport::new(Arc::new(client))));
         let mut req = Request::new("https://localhost".parse().unwrap(), Method::Get);
 
         let err = policy
@@ -248,7 +246,7 @@ mod tests {
             }
             .boxed()
         }));
-        let transport = Arc::new(TransportPolicy::new(TransportOptions::new(client)));
+        let transport = Arc::new(TransportPolicy::new(Transport::new(client)));
 
         let mut handles = vec![];
         for _ in 0..4 {